refactor(metrics): split primitives into per-type modules + drop MetricsAdapter per ADR-0046
ADR-0046 §Decision specified a flat per-primitive module layout
(counter / gauge / histogram / registry / labels / naming / filter /
prometheus / eventbus / error / prelude) and the deletion of the
TelemetryAdapter / MetricsAdapter bridge. The earlier merge commits
(2794744a / 33cdd025 / b804048f) absorbed nebula-telemetry but kept
Counter/Gauge/Histogram inside registry.rs and renamed the bridge to
MetricsAdapter, leaving daily friction the ADR set out to remove. This
finishes the spec.

- Move Counter -> counter.rs, Gauge -> gauge.rs, Histogram /
  HistogramSnapshot / DEFAULT_BUCKETS -> histogram.rs.
- Slim registry.rs to MetricsRegistry + MetricSeries + the registry
  tests; expose now_ms as pub(crate) for the per-primitive modules.
- Replace MetricsAdapter with a free function record_eventbus_stats(
  &MetricsRegistry, &EventBusStats) -> MetricsResult<()> in eventbus.rs.
  Typed canonical-name accessors disappear; callers reach the same
  metrics through MetricsRegistry::counter / gauge / histogram with
  NEBULA_* constants from naming::*.
- Update tests/integration.rs and both examples (cardinality_guard,
  prometheus_export) to use the registry directly.
- lib.rs adopts the ADR-0046 grouping (primitives / policy / export /
  instrumentation / error / prelude) with private submodules and a flat
  pub use surface; naming stays pub mod for downstream users that
  import nebula_metrics::naming::*. pub use naming::* preserves the
  root-level NEBULA_* re-exports nebula-engine relies on.
- Sweep stale references: crates/api/README.md, crates/log/README.md,
  crates/log/docs/Integration.md, .ai-factory/rules/base.md no longer
  mention nebula-telemetry / TelemetryAdapter.

Verification (worktree refactor/metrics-telemetry-merge):
- cargo check --workspace --all-targets clean
- cargo clippy --workspace --all-targets -- -D warnings clean
- cargo fmt --all -- --check clean
- cargo nextest run -p nebula-metrics: 70/70 pass
- cargo test -p nebula-metrics --doc: 10/10 pass + 1 ignored
- cargo nextest run --workspace clean
- cargo test --doc --workspace clean
- cargo deny check clean
