--- crates/metrics/src/adapter.rs
@@ -1,425 +1,0 @@
-//! Adapter over `MetricsRegistry` that records using standard `nebula_*` names.
-//!
-//! Use this when you want to record metrics with the unified naming convention
-//! without changing the underlying telemetry registry.
-
-use std::sync::Arc;
-
-use crate::{
-    labels::LabelSet,
-    registry::{Counter, Gauge, Histogram, MetricsRegistry},
-};
-use nebula_eventbus::EventBusStats;
-
-use crate::{MetricsResult, filter::LabelAllowlist, naming::*};
-
-/// Adapter that exposes telemetry metrics under standard `nebula_*` names.
-///
-/// Holds an `Arc<MetricsRegistry>` and provides accessors that return
-/// counters/gauges/histograms with the canonical names. Engine and runtime
-/// can use this to record metrics that are ready for Prometheus/OTLP export.
-///
-/// An optional [`LabelAllowlist`] can be configured via
-/// [`MetricsAdapter::with_allowlist`] to strip high-cardinality label keys
-/// before they reach the registry.
-#[derive(Clone, Debug)]
-pub struct MetricsAdapter {
-    registry: Arc<MetricsRegistry>,
-    allowlist: LabelAllowlist,
-}
-
-impl MetricsAdapter {
-    /// Create an adapter over the given telemetry registry.
-    ///
-    /// By default all labels are passed through. Use
-    /// [`MetricsAdapter::with_allowlist`] to restrict which keys are stored.
-    #[must_use]
-    pub fn new(registry: Arc<MetricsRegistry>) -> Self {
-        Self {
-            registry,
-            allowlist: LabelAllowlist::all(),
-        }
-    }
-
-    /// Attach a [`LabelAllowlist`] that filters label keys on every labeled
-    /// accessor call.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use std::sync::Arc;
-    ///
-    /// use nebula_metrics::{LabelAllowlist, MetricsAdapter};
-    /// use nebula_metrics::MetricsRegistry;
-    ///
-    /// let reg = Arc::new(MetricsRegistry::new());
-    /// let adapter = MetricsAdapter::new(Arc::clone(&reg)).with_allowlist(LabelAllowlist::only([
-    ///     "action_type",
-    ///     "status",
-    ///     "trigger_type",
-    /// ]));
-    /// ```
-    #[must_use]
-    pub fn with_allowlist(mut self, allowlist: LabelAllowlist) -> Self {
-        self.allowlist = allowlist;
-        self
-    }
-
-    /// Apply the configured [`LabelAllowlist`] to `labels`.
-    ///
-    /// Use this to build a safe [`LabelSet`] before calling labeled accessors
-    /// when the incoming label set may contain high-cardinality keys.
-    #[must_use]
-    pub fn filter_labels(&self, labels: &LabelSet) -> LabelSet {
-        self.allowlist.apply(labels, self.registry.interner())
-    }
-
-    /// Underlying registry for custom or legacy metric names.
-    #[must_use]
-    pub fn registry(&self) -> &MetricsRegistry {
-        &self.registry
-    }
-
-    // ---------- Workflow (engine) ----------
-
-    /// Counter: workflow executions started.
-    pub fn workflow_executions_started_total(&self) -> MetricsResult<Counter> {
-        self.registry
-            .counter(NEBULA_WORKFLOW_EXECUTIONS_STARTED_TOTAL)
-    }
-
-    /// Counter: workflow executions completed successfully.
-    pub fn workflow_executions_completed_total(&self) -> MetricsResult<Counter> {
-        self.registry
-            .counter(NEBULA_WORKFLOW_EXECUTIONS_COMPLETED_TOTAL)
-    }
-
-    /// Counter: workflow executions failed.
-    pub fn workflow_executions_failed_total(&self) -> MetricsResult<Counter> {
-        self.registry
-            .counter(NEBULA_WORKFLOW_EXECUTIONS_FAILED_TOTAL)
-    }
-
-    /// Histogram: workflow execution duration in seconds.
-    pub fn workflow_execution_duration_seconds(&self) -> MetricsResult<Histogram> {
-        self.registry
-            .histogram(NEBULA_WORKFLOW_EXECUTION_DURATION_SECONDS)
-    }
-
-    // ---------- Action (runtime) ----------
-
-    /// Counter: action executions (success + failure).
-    pub fn action_executions_total(&self) -> MetricsResult<Counter> {
-        self.registry.counter(NEBULA_ACTION_EXECUTIONS_TOTAL)
-    }
-
-    /// Counter: action failures.
-    pub fn action_failures_total(&self) -> MetricsResult<Counter> {
-        self.registry.counter(NEBULA_ACTION_FAILURES_TOTAL)
-    }
-
-    /// Histogram: action execution duration in seconds.
-    pub fn action_duration_seconds(&self) -> MetricsResult<Histogram> {
-        self.registry.histogram(NEBULA_ACTION_DURATION_SECONDS)
-    }
-
-    // ---------- Generic access (for domain-specific names) ----------
-
-    /// Get or create a counter by name. Prefer the typed accessors above when available.
-    pub fn counter(&self, name: &str) -> MetricsResult<Counter> {
-        self.registry.counter(name)
-    }
-
-    /// Get or create a gauge by name.
-    pub fn gauge(&self, name: &str) -> MetricsResult<Gauge> {
-        self.registry.gauge(name)
-    }
-
-    /// Get or create a histogram by name.
-    pub fn histogram(&self, name: &str) -> MetricsResult<Histogram> {
-        self.registry.histogram(name)
-    }
-
-    // ---------- Labeled variants of canonical action metrics ----------
-
-    /// Counter: action executions with label dimensions (e.g. `action_type`).
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::sync::Arc;
-    ///
-    /// use nebula_metrics::adapter::MetricsAdapter;
-    /// use nebula_metrics::MetricsRegistry;
-    ///
-    /// let reg = Arc::new(MetricsRegistry::new());
-    /// let adapter = MetricsAdapter::new(Arc::clone(&reg));
-    /// let labels = reg.interner().label_set(&[("action_type", "http.request")]);
-    /// adapter.action_executions_labeled(&labels).unwrap().inc();
-    /// ```
-    pub fn action_executions_labeled(&self, labels: &LabelSet) -> MetricsResult<Counter> {
-        let labels = self.filter_labels(labels);
-        self.registry
-            .counter_labeled(NEBULA_ACTION_EXECUTIONS_TOTAL, &labels)
-    }
-
-    /// Counter: action failures with label dimensions.
-    pub fn action_failures_labeled(&self, labels: &LabelSet) -> MetricsResult<Counter> {
-        let labels = self.filter_labels(labels);
-        self.registry
-            .counter_labeled(NEBULA_ACTION_FAILURES_TOTAL, &labels)
-    }
-
-    /// Histogram: action execution duration with label dimensions.
-    pub fn action_duration_labeled(&self, labels: &LabelSet) -> MetricsResult<Histogram> {
-        let labels = self.filter_labels(labels);
-        self.registry
-            .histogram_labeled(NEBULA_ACTION_DURATION_SECONDS, &labels)
-    }
-
-    /// Counter: workflow executions started with label dimensions.
-    pub fn workflow_executions_started_labeled(&self, labels: &LabelSet) -> MetricsResult<Counter> {
-        let labels = self.filter_labels(labels);
-        self.registry
-            .counter_labeled(NEBULA_WORKFLOW_EXECUTIONS_STARTED_TOTAL, &labels)
-    }
-
-    /// Counter: workflow executions completed with label dimensions.
-    pub fn workflow_executions_completed_labeled(
-        &self,
-        labels: &LabelSet,
-    ) -> MetricsResult<Counter> {
-        let labels = self.filter_labels(labels);
-        self.registry
-            .counter_labeled(NEBULA_WORKFLOW_EXECUTIONS_COMPLETED_TOTAL, &labels)
-    }
-
-    /// Counter: workflow executions failed with label dimensions.
-    pub fn workflow_executions_failed_labeled(&self, labels: &LabelSet) -> MetricsResult<Counter> {
-        let labels = self.filter_labels(labels);
-        self.registry
-            .counter_labeled(NEBULA_WORKFLOW_EXECUTIONS_FAILED_TOTAL, &labels)
-    }
-
-    /// Access the underlying label interner to build [`LabelSet`]s.
-    #[must_use]
-    pub fn interner(&self) -> &crate::labels::LabelInterner {
-        self.registry.interner()
-    }
-
-    // ---------- EventBus snapshots ----------
-
-    /// Gauge: eventbus sent events snapshot.
-    pub fn eventbus_sent(&self) -> MetricsResult<Gauge> {
-        self.registry.gauge(NEBULA_EVENTBUS_SENT)
-    }
-
-    /// Gauge: eventbus dropped events snapshot.
-    pub fn eventbus_dropped(&self) -> MetricsResult<Gauge> {
-        self.registry.gauge(NEBULA_EVENTBUS_DROPPED)
-    }
-
-    /// Gauge: eventbus active subscriber snapshot.
-    pub fn eventbus_subscribers(&self) -> MetricsResult<Gauge> {
-        self.registry.gauge(NEBULA_EVENTBUS_SUBSCRIBERS)
-    }
-
-    /// Gauge: eventbus drop ratio snapshot in parts-per-million (`0..=1_000_000`).
-    pub fn eventbus_drop_ratio_ppm(&self) -> MetricsResult<Gauge> {
-        self.registry.gauge(NEBULA_EVENTBUS_DROP_RATIO_PPM)
-    }
-
-    /// Records an [`EventBusStats`] snapshot under standard `nebula_eventbus_*` gauges.
-    pub fn record_eventbus_stats(&self, stats: &EventBusStats) -> MetricsResult<()> {
-        self.eventbus_sent()?
-            .set(clamp_u64_to_i64(stats.sent_count));
-        self.eventbus_dropped()?
-            .set(clamp_u64_to_i64(stats.dropped_count));
-        self.eventbus_subscribers()?
-            .set(clamp_usize_to_i64(stats.subscriber_count));
-
-        let ppm = (stats.drop_ratio() * 1_000_000.0).round();
-        let ppm = if ppm.is_finite() {
-            ppm.clamp(0.0, i64::MAX as f64) as i64
-        } else {
-            0
-        };
-        self.eventbus_drop_ratio_ppm()?.set(ppm);
-        Ok(())
-    }
-}
-
-fn clamp_u64_to_i64(value: u64) -> i64 {
-    i64::try_from(value).unwrap_or(i64::MAX)
-}
-
-fn clamp_usize_to_i64(value: usize) -> i64 {
-    i64::try_from(value).unwrap_or(i64::MAX)
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-
-    use crate::registry::MetricsRegistry;
-    use nebula_eventbus::EventBusStats;
-
-    use super::MetricsAdapter;
-    use crate::LabelAllowlist;
-
-    #[test]
-    fn adapter_records_under_standard_names() {
-        let registry = Arc::new(MetricsRegistry::new());
-        let adapter = MetricsAdapter::new(Arc::clone(&registry));
-
-        adapter.workflow_executions_started_total().unwrap().inc();
-        adapter.workflow_executions_started_total().unwrap().inc();
-        adapter.action_executions_total().unwrap().inc();
-        adapter.action_failures_total().unwrap().inc();
-
-        assert_eq!(
-            adapter.workflow_executions_started_total().unwrap().get(),
-            2
-        );
-        assert_eq!(adapter.action_executions_total().unwrap().get(), 1);
-        assert_eq!(adapter.action_failures_total().unwrap().get(), 1);
-    }
-
-    #[test]
-    fn adapter_records_eventbus_snapshot_metrics() {
-        let registry = Arc::new(MetricsRegistry::new());
-        let adapter = MetricsAdapter::new(Arc::clone(&registry));
-
-        let stats = EventBusStats {
-            sent_count: 75,
-            dropped_count: 25,
-            subscriber_count: 3,
-        };
-
-        adapter.record_eventbus_stats(&stats).unwrap();
-
-        assert_eq!(adapter.eventbus_sent().unwrap().get(), 75);
-        assert_eq!(adapter.eventbus_dropped().unwrap().get(), 25);
-        assert_eq!(adapter.eventbus_subscribers().unwrap().get(), 3);
-        assert_eq!(adapter.eventbus_drop_ratio_ppm().unwrap().get(), 250_000);
-    }
-
-    #[test]
-    fn record_eventbus_stats_handles_zero_totals_and_default_values() {
-        let adapter = MetricsAdapter::new(Arc::new(MetricsRegistry::new()));
-
-        let stats = EventBusStats {
-            sent_count: 0,
-            dropped_count: 0,
-            subscriber_count: 0,
-        };
-        tracing::debug!(
-            "record_eventbus_stats test: sent={} dropped={} ppm_expected={}",
-            stats.sent_count,
-            stats.dropped_count,
-            0
-        );
-        adapter.record_eventbus_stats(&stats).unwrap();
-        assert_eq!(adapter.eventbus_sent().unwrap().get(), 0);
-        assert_eq!(adapter.eventbus_dropped().unwrap().get(), 0);
-        assert_eq!(adapter.eventbus_subscribers().unwrap().get(), 0);
-        assert_eq!(adapter.eventbus_drop_ratio_ppm().unwrap().get(), 0);
-
-        let default_stats = EventBusStats::default();
-        tracing::debug!(
-            "record_eventbus_stats test: sent={} dropped={} ppm_expected={}",
-            default_stats.sent_count,
-            default_stats.dropped_count,
-            0
-        );
-        adapter.record_eventbus_stats(&default_stats).unwrap();
-        assert_eq!(adapter.eventbus_sent().unwrap().get(), 0);
-        assert_eq!(adapter.eventbus_dropped().unwrap().get(), 0);
-        assert_eq!(adapter.eventbus_subscribers().unwrap().get(), 0);
-        assert_eq!(adapter.eventbus_drop_ratio_ppm().unwrap().get(), 0);
-    }
-
-    #[test]
-    fn record_eventbus_stats_handles_full_drop_ratio_and_rounding() {
-        let adapter = MetricsAdapter::new(Arc::new(MetricsRegistry::new()));
-
-        let full_drop = EventBusStats {
-            sent_count: 1_000_000,
-            dropped_count: 1_000_000,
-            subscriber_count: 42,
-        };
-        tracing::debug!(
-            "record_eventbus_stats test: sent={} dropped={} ppm_expected={}",
-            full_drop.sent_count,
-            full_drop.dropped_count,
-            1_000_000
-        );
-        adapter.record_eventbus_stats(&full_drop).unwrap();
-        assert_eq!(adapter.eventbus_drop_ratio_ppm().unwrap().get(), 500_000);
-
-        let fractional = EventBusStats {
-            sent_count: 3,
-            dropped_count: 1,
-            subscriber_count: 1,
-        };
-        tracing::debug!(
-            "record_eventbus_stats test: sent={} dropped={} ppm_expected={}",
-            fractional.sent_count,
-            fractional.dropped_count,
-            250_000
-        );
-        adapter.record_eventbus_stats(&fractional).unwrap();
-        assert_eq!(adapter.eventbus_drop_ratio_ppm().unwrap().get(), 250_000);
-    }
-
-    #[test]
-    fn record_eventbus_stats_clamps_large_values_to_i64_max() {
-        let adapter = MetricsAdapter::new(Arc::new(MetricsRegistry::new()));
-
-        let stats = EventBusStats {
-            sent_count: u64::MAX,
-            dropped_count: 0,
-            subscriber_count: usize::MAX,
-        };
-        tracing::debug!(
-            "record_eventbus_stats test: sent={} dropped={} ppm_expected={}",
-            stats.sent_count,
-            stats.dropped_count,
-            0
-        );
-        adapter.record_eventbus_stats(&stats).unwrap();
-
-        assert_eq!(adapter.eventbus_sent().unwrap().get(), i64::MAX);
-        assert_eq!(adapter.eventbus_subscribers().unwrap().get(), i64::MAX);
-        assert_eq!(adapter.eventbus_drop_ratio_ppm().unwrap().get(), 0);
-    }
-
-    #[test]
-    fn labeled_accessors_apply_allowlist_before_recording() {
-        let registry = Arc::new(MetricsRegistry::new());
-        let adapter = MetricsAdapter::new(Arc::clone(&registry))
-            .with_allowlist(LabelAllowlist::only(["action_type"]));
-        let labels = registry.interner().label_set(&[
-            ("action_type", "http.request"),
-            ("execution_id", "exec-123"),
-        ]);
-
-        adapter.action_executions_labeled(&labels).unwrap().inc();
-
-        let snapshots = registry.snapshot_counters();
-        let (key, counter) = snapshots
-            .iter()
-            .find(|(key, _)| {
-                registry.interner().resolve(key.name) == "nebula_action_executions_total"
-            })
-            .expect("counter should be recorded");
-        assert_eq!(counter.get(), 1);
-
-        let labels = key.labels.resolve(registry.interner());
-        assert_eq!(
-            labels,
-            vec![("action_type", "http.request")],
-            "high-cardinality execution_id must be filtered out",
-        );
-    }
-}
