+++ crates/metrics/tests/integration.rs
@@ -1,26 +1,22 @@
-use std::sync::Arc;
-
 use nebula_metrics::{
-    MetricsAdapter, MetricsRegistry,
+    MetricsRegistry,
     naming::{NEBULA_RESOURCE_ACQUIRE_WAIT_DURATION_SECONDS, NEBULA_RESOURCE_CREATE_TOTAL},
 };
 
 #[test]
-fn telemetry_adapter_resource_metrics_round_trip_via_generic_accessors() {
-    let registry = Arc::new(MetricsRegistry::new());
-    let adapter = MetricsAdapter::new(Arc::clone(&registry));
+fn registry_records_resource_metrics_via_canonical_names() {
+    let registry = MetricsRegistry::new();
+
+    tracing::debug!("incrementing {NEBULA_RESOURCE_CREATE_TOTAL} via registry");
+    registry
+        .counter(NEBULA_RESOURCE_CREATE_TOTAL)
+        .unwrap()
+        .inc();
 
     tracing::debug!(
-        "incrementing {} via generic adapter",
-        NEBULA_RESOURCE_CREATE_TOTAL
+        "observing {NEBULA_RESOURCE_ACQUIRE_WAIT_DURATION_SECONDS} via registry with sample=0.5"
     );
-    adapter.counter(NEBULA_RESOURCE_CREATE_TOTAL).unwrap().inc();
-
-    tracing::debug!(
-        "observing {} via generic adapter with sample=0.5",
-        NEBULA_RESOURCE_ACQUIRE_WAIT_DURATION_SECONDS
-    );
-    adapter
+    registry
         .histogram(NEBULA_RESOURCE_ACQUIRE_WAIT_DURATION_SECONDS)
         .unwrap()
         .observe(0.5);
